+++ core/libs/database/coredb/coredb.h
@@ -760,7 +760,7 @@ public:
     /**
      * Updates the manualOrder field for the item.
      */
-    void setItemManualOrder(qlonglong imageID, int value);
+    void setItemManualOrder(qlonglong imageID, qlonglong value);
 
     /**
      * Rename the item.
